+++ packages/components/src/components/Tabs/Tabs.tsx
@@ -3,16 +3,20 @@ import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
 import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
 import clsx from "clsx";
-import type { PropsWithChildren } from "react";
+import type { PropsWithChildren, ReactNode } from "react";
 import { useState } from "react";
 import * as Aria from "react-aria-components";
 import styles from "./Tabs.module.scss";
+import { FallbackTab } from "@/components/Tabs/components/FallbackTab";
 
 export interface TabsProps
   extends
     Omit<Aria.TabsProps, "children">,
     PropsWithChildren,
-    FlowComponentProps {}
+    FlowComponentProps {
+  /* custom fallback view for not found tabs */
+  tabNotFoundView?: ReactNode;
+}
 
 /** @flr-generate all */
 export const Tabs = flowComponent("Tabs", (props) => {
@@ -24,6 +28,7 @@ export const Tabs = flowComponent("Tabs", (props) => {
     disabledKeys,
     ref,
     onSelectionChange,
+    tabNotFoundView,
     ...rest
   } = props;
 
@@ -62,6 +67,7 @@ export const Tabs = flowComponent("Tabs", (props) => {
           disabledKeys={disabledKeys}
         />
         <TunnelExit id="Panels" />
+        <FallbackTab tabNotFoundView={tabNotFoundView} />
       </Aria.Tabs>
     </TunnelProvider>
   );