+++ packages/remote-elements/src/auto-generated/RemoteRangeCalendarElement.ts
@@ -40,6 +40,7 @@ export class RemoteRangeCalendarElement extends FlowRemoteElement<RemoteRangeCal
       translate: {},
       value: {},
       visibleDuration: {},
+      withDatePickerPresets: {},
     };
   }